requirements.txt
# CERBERUS - Sistema de Análise Criminal
# Dependências principais

# Data validation and models
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# String matching and normalization
unidecode>=1.3.0
python-Levenshtein>=0.21.0

# Configuration management
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
